@@ -3,9 +3,10 @@
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
+  "type": "module",
   "engines": {
-    "node": "14.16.0",
-    "npm": "6.14.11"
+    "node": "20.6.1",
+    "npm": "9.8.1"
   },
   "scripts": {
     "start": "node index.js",
@@ -14,13 +15,16 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "bcrypt": "^5.0.1",
+    "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.0",
     "connect-pg-simple": "^7.0.0",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
+    "express-rate-limit": "^7.1.3",
     "express-session": "^1.17.3",
+    "file-type": "^18.6.0",
+    "fs-extra": "^11.1.1",
     "helmet": "^5.1.0",
     "moment": "^2.29.1",
     "multer": "^1.4.5-lts.1",
@@ -28,8 +32,10 @@
     "nodemon": "^2.0.16",
     "pg": "^8.6.0",
     "pg-range": "^1.1.0",
+    "qrcode": "^1.5.3",
     "sharp": "^0.30.7",
     "socket.io": "^4.5.1",
+    "speakeasy": "^2.0.0",
     "yup": "^0.32.11"
   }
 }
