@@ -1,5 +1,9 @@
-const Pool = require("pg").Pool;
-require("dotenv").config();
+import pg from "pg";
+const { Pool } = pg;
+import { URL } from "url";
+import dotenv from "dotenv";
+
+dotenv.config();
 //podaci za login na bazu
 const devConfig = `postgres://${process.env.PG_USER}:${process.env.PG_PASSWORD}@${process.env.PG_HOST}:${process.env.PG_PORT}/${process.env.PG_DATABASE}`;
 const proConfig = process.env.DATABASE_URL;
@@ -12,4 +16,4 @@ const pool = new Pool({
       : false,
 });
 
-module.exports = pool;
+export default pool;
