@@ -1,20 +1,23 @@
-const express = require("express");
-const cors = require("cors");
-const { Server } = require("socket.io");
+import express from "express";
+import { Server } from "socket.io";
+import helmet from "helmet";
+import cors from "cors";
+import authRouter from "./routers/authRouter.js";
+import session from "express-session";
+import connectPg from "connect-pg-simple";
+import pool from "./db.js";
+import { createServer } from "http";
+import dotenv from "dotenv";
+import { fileURLToPath } from "url";
+import { dirname, join } from "path";
+import pg from "pg";
 const app = express();
-const path = require("path");
-const pool = require("./db");
-const PORT = process.env.PORT || 3001;
-const helmet = require("helmet");
-const session = require("express-session");
-const connectPg = require("connect-pg-simple");
-const authRouter = require("./routers/authRouter");
 const pgSession = connectPg(session);
-const server = require("http").createServer(app);
-const nodemailer = require("nodemailer");
-const moment = require("moment");
-require("dotenv").config();
+const server = createServer(app);
+const PORT = process.env.PORT || 3001;
 //middleware
+dotenv.config();
+app.set("trust proxy", 1);
 const io = new Server(server, {
   cors: {
     origin: "http://localhost:3000",
@@ -34,9 +37,9 @@ app.use(
     directives: {
       "default-src": ["'self'"],
       "connect-src": ["'self'", "blob:", "https:"],
-      "img-src": ["'self'", "data:", "https:"],
+      "img-src": ["'self'", "data:", "https:", "https: data: blob:"],
       "style-src": ["'self'", "data:", "https:", "'unsafe-inline'"],
-      "script-src": ["'unsafe-inline'", "'self'", "https:"],
+      "script-src": ["'unsafe-inline'", "'self'", "https:", "'unsafe-eval'"],
       "object-src": ["'none'"],
     },
   })
@@ -70,21 +73,33 @@ app.use(
     },
   })
 );
-app.use("/uploads", (req, res, next) => {
-  if (req.session.user && req.session.user.username) {
-    // user postoji i sesija nastavi dalje.
+const authorizationMiddleware = (req, res, next) => {
+  // Extract the username from the URL, assuming the URL format is '/uploads/usersprofile/{username}/...'
+  const requestedUsername = req.path.split("/")[1];
+  console.log(requestedUsername);
+  // Check if the user is authenticated and if the authenticated user's username matches the requested username
+  if (req.session.user && req.session.user.username === requestedUsername) {
     next();
   } else {
-    // No user object found, terminate the pipeline with .end().
+    res.status(403).send("Forbidden");
+  }
+};
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+app.use(express.static(join(__dirname, "client/build")));
+
+app.use("/uploads", (req, res, next) => {
+  if (req.session.user && req.session.user.username) {
+    next();
+  } else {
     res.status(401).end();
   }
 });
-app.use(
-  "/uploads",
-  express.static(path.join(__dirname, "./uploads/usersprofile"))
-);
-app.use(express.static(path.join(__dirname, "client/build")));
-//rute
+
+// Authorization middleware for user profile pictures
+app.use("/uploads/usersprofile", authorizationMiddleware);
+// Serve all static files from the /uploads directory (including nested directories)
+app.use("/uploads", express.static(join(__dirname, "./uploads")));
 if (process.env.NODE_ENV === "production") {
   app.use(express.static(path.join(__dirname, "client/build")));
 }
@@ -92,12 +107,17 @@ if (process.env.NODE_ENV === "production") {
 
 app.use("/auth", authRouter);
 //brisanje rezervacija
-
+io.on("connect", (socket) => {
+  socket.on("error", (err) => {
+    console.error(`Socket error: ${err}`);
+  });
+});
 //umetanje rezervacije i slanje maila kod poslanog zahtjeva za rezervaciju
 
 app.get("*", (req, res) => {
-  res.sendFile(path.join(__dirname, "client/build/index.html"));
+  res.sendFile(join(__dirname, "client/build/index.html"));
 });
+
 io.on("connect", (socket) => {});
 //listen na portu 3001
 app.listen(PORT, () => {
